When no scope is provided to the Strategy, do not request any
This is to support implicit scopes. When we make the OIDC request, the url should be:
https://issuer.com/authorize?response_type=code&client_id=myapp-id&redirect_uri=myapp/login
The library, by default, follows what the spec says, that no token should be given without the 'openid' scope.
However, our issuer service supports an implicit set of scopes that will be used in the event that no
  scopes are requested. Therefore, we have to delete the 'scope' property entirely.
If we do not, the request url will have '&scope=' attached to the end, which is explicitly
  requesting zero scopes (to which the issuing service is happy to oblige).